add node server
