+++ NodeServer/Readme.txt
@@ -1,0 +1,3 @@
+접속 서버 루트경로 : https://golden-finder.firebaseapp.com
+
+현재까지 만든 하위경로 : /api
